More board drawing fixes
- Ensure the QGraphicsView has no border, so that everything fits
- Use doubles for squareSize, so that there's less wasted space
- Assorted coordinate fixes

+++ src/gatter.cpp
@@ -115,7 +115,7 @@ Gatter::~Gatter()
  /**
   * Calculates the gatter intersections and hoshis position
   **/
-void Gatter::resize(int offsetX, int offsetY, int square_size)
+void Gatter::resize(int offsetX, int offsetY, double square_size)
 {
 	int i,j;
 	QGraphicsEllipseItem *e;
